+++ plugins/gs-common/src/model/capiv1beta1/key.ts
@@ -1,3 +1,5 @@
+export const clusterApiVersion = 'cluster.x-k8s.io/v1beta1';
+
 export const clusterGVK = {
   apiVersion: 'v1beta1',
   group: 'cluster.x-k8s.io',