+++ plugins/gs/src/components/InstallationsSelector/InstallationsSelector.tsx
@@ -11,7 +11,7 @@ import {
   makeStyles,
 } from '@material-ui/core';
 import classnames from 'classnames';
-import { InstallationStatus } from '../hooks/useInstallationsStatuses';
+import type { InstallationStatus } from '../hooks/useInstallationsStatuses';
 
 const useStyles = makeStyles(theme => ({
   formControl: {