+++ plugins/gs/src/components/UI/GitOpsUILink/index.ts
@@ -1,0 +1,1 @@
+export { GitOpsUILink } from './GitOpsUILink';