+++ plugins/gs/src/components/clusters/ClusterDetails/index.ts
@@ -1,0 +1,1 @@
+export { ClusterDetails } from './ClusterDetails';