+++ plugins/gs/src/components/clusters/ClustersPage/index.ts
@@ -1,0 +1,1 @@
+export { ClustersPage } from './ClustersPage';