+++ plugins/gs/src/plugin.ts
@@ -3,28 +3,37 @@ import {
   createRoutableExtension,
 } from '@backstage/core-plugin-api';
 
-import { entityDeploymentsRouteRef, rootRouteRef } from './routes';
+import {
+  clustersRouteRef,
+  entityDeploymentsRouteRef,
+  rootRouteRef,
+} from './routes';
 
 export const gsPlugin = createPlugin({
   id: 'gs',
   routes: {
     root: rootRouteRef,
+    clustersPage: clustersRouteRef,
     entityContent: entityDeploymentsRouteRef,
   },
 });
 
-export const GSPluginPage = gsPlugin.provide(
+export const GSClustersPage = gsPlugin.provide(
   createRoutableExtension({
-    name: 'GSPluginPage',
-    component: () => import('./components/PluginPage').then(m => m.PluginPage),
-    mountPoint: rootRouteRef,
+    name: 'GSClustersPage',
+    component: () =>
+      import('./components/clusters/ClustersPage').then(m => m.ClustersPage),
+    mountPoint: clustersRouteRef,
   }),
 );
 
 export const EntityGSDeploymentsContent = gsPlugin.provide(
   createRoutableExtension({
     name: 'EntityGSDeploymentsContent',
-    component: () => import('./components/Router').then(m => m.Router),
+    component: () =>
+      import('./components/deployments/EntityDeploymentsContent').then(
+        m => m.EntityDeploymentsContent,
+      ),
     mountPoint: entityDeploymentsRouteRef,
   }),
 );