+++ plugins/gs/src/routes.ts
@@ -1,15 +1,13 @@
-import { createRouteRef, createSubRouteRef } from '@backstage/core-plugin-api';
+import { createRouteRef } from '@backstage/core-plugin-api';
 
 export const rootRouteRef = createRouteRef({
   id: 'gs',
 });
 
-export const entityDeploymentsRouteRef = createRouteRef({
-  id: 'gs-deployments',
+export const clustersRouteRef = createRouteRef({
+  id: 'gs-clusters',
 });
 
-export const entityDeploymentDetailsRouteRef = createSubRouteRef({
-  id: 'gs-deployments/details',
-  path: '/:deploymentInstallation/:deploymentNamespace/:deploymentName',
-  parent: entityDeploymentsRouteRef,
+export const entityDeploymentsRouteRef = createRouteRef({
+  id: 'gs-deployments',
 });